machines: Pull volume creation content into separate component
