+++ test/selenium/selenium-machines-disks.py
@@ -36,7 +36,7 @@ class MachinesDisksTestSuite(MachinesLib):
         self.click(self.wait_css('#vm-{}-disks-adddisk'.format(name), cond=clickable))
         self.select_by_value(self.wait_css('#vm-{}-disks-adddisk-new-select-pool'.format(name)), pool[2])
         self.send_keys(self.wait_css('#vm-{}-disks-adddisk-new-name'.format(name)), 'raw2disk_' + MachinesLib.random_string())
-        self.select_by_value(self.wait_css('#vm-{}-disks-adddisk-new-diskfileformat'.format(name)), 'raw')
+        self.select_by_value(self.wait_css('#vm-{}-disks-adddisk-new-format'.format(name)), 'raw')
         self.click(self.wait_css('#vm-{}-disks-adddisk-dialog-add'.format(name), cond=clickable))
         self.wait_dialog_disappear()
         self.wait_css('#vm-{}-disks-vdb-device'.format(name))
@@ -80,7 +80,7 @@ class MachinesDisksTestSuite(MachinesLib):
         self.click(self.wait_css('#vm-{}-disks-adddisk'.format(name), cond=clickable))
         self.select_by_value(self.wait_css('#vm-{}-disks-adddisk-new-select-pool'.format(name)), pool[1])
         self.send_keys(self.wait_css('#vm-{}-disks-adddisk-new-name'.format(name)), 'raw2disk_' + MachinesLib.random_string())
-        self.select_by_value(self.wait_css('#vm-{}-disks-adddisk-new-diskfileformat'.format(name)), 'raw')
+        self.select_by_value(self.wait_css('#vm-{}-disks-adddisk-new-format'.format(name)), 'raw')
         self.click(self.wait_css('#vm-{}-disks-adddisk-dialog-add'.format(name), cond=clickable))
         self.wait_dialog_disappear()
         self.wait_css('#vm-{}-disks-vdb-device'.format(name))
